refactor: BindingAdapter를 기능별로 분리
